feat(shared-prompts): restructure unified prompt request and renderer
Made-with: Cursor
