+++ sharedPrompts/src/main/java/org/example/sharedprompts/domain/prompt/common/enums/RequestType.java
@@ -1,0 +1,12 @@
+package org.example.sharedprompts.domain.prompt.common.enums;
+
+/**
+ * 통합 프롬프트 생성 요청 타입 디스크리미네이터.
+ */
+public enum RequestType {
+
+    SIMPLE,
+    EXTRACTION,
+    ADVANCED
+}
+
